add regression test for #12546 (#12556)
This is a minimized e2e regression test for #12546. The apparent symptom
there was e2e matmul tests with data tiling were failing. This test
narrows it down to the `tensor.pack` of the RHS specifically (the one
side that gets transposed) and for very specific outer-product shapes
(here RHS is 1x9).