Be consistent about the use of IREE Bazel macros (#12645)
We have some wrapper macros for various rules, mostly to add includes,
but usage was not consistent. This still worked because includes are
inherited transitively, but it only worked by accident basically.

Classifying things this way is also helpful for differentiating compiler
and runtime in bazel query and such (don't need to analyze the whole dep
tree). It could also provide hook points to guard against
cross-component dependencies.