[LLVMCPU] Make pack fusion dispatches select config based on producers. (#12590)
The pack op only impact the store position of destination buffers. We
should select the config based on producers for fusion cases.

The TileAndDistribute pass is updated to handle that there are encoding ops
but the pack op is not a root op.

The `getVectorSizes` is also updated because the tiling sizes and the shape
are not always identical. In the pack op fusion case, the actual shapes could
be multiples of tiling sizes. They are scaled to valid numbers when they are static.