Avoid FunctionalStyleTransformOpTrait where unnecessary to improve u… (#12681)
…sability

This PR simplifies most transforms that return the same handle that they
were passed without modifying the op itself. The chaining behavior
results in too much cognitive overhead when manipulating transform
scripts as handle names need to constantly be updated as soon as a
transform is disabled or moved. Instead, avoid consuming the handle
wherever possible.