Add option to TileAndDistribute to not emit scf.for distribution loop (#12784)
We currently distribute in a way that each thread gets one tile. This
means we don't need to emit the disatribution loops. Add an option to
avoid emiting those scf.for, this can be used by backends to simplify
the codegen and not rely on britle pass RemoveTrivialLoops. This also
expand affine_min canonicalization to not rely on distributed for loops.