[spirv] Tile reduction to avoid code bloat (#12922)
Previously we only tile reduction dimensions for multiples of 4 when
going down the vectorization pipeline. This causes code bloat and long
compilation time for odd dimension sizes which gets fully unrolled.
Improve the logic to try to find the smallest prime number for tiling
such cases too, before we have the solution of general packing and
padding.