[spirv] Fix CodeGen for no surrounding distributed workgroup loops (#12940)
https://github.com/openxla/iree/pull/12800 converts SPIR-V
CodeGen to avoid emitting distributed loops for workgroups.
This means we cannot deduce the number of workgroups from
the loop bounds anymore. To simply `affine.min` ops involving
workgroup ID ops, we need to query the number of workgroups
separately and feed that information to folding patterns.

This commit wires the above to connect the dots to make the
information flow. Additionally we convert the FastSlowPathPass
in SPIR-V to not anchor on surrounding distributed loops.

---------

Co-authored-by: Lei Zhang <antiagainst@google.com>