[StableHLO][NFC] Split linalg lowering tests (#13059)
Factor out pointwise and dot prod op tests from the main test file. This
matches the source file organization of the conversion patterns.

The test cases are not modified. The only difference is that now we do
not run with the 'enable-primitive-ops' option on tests files that are
not affected by it.

Issue: https://github.com/openxla/iree/issues/12678