`ukernel.generic_raw` op passing only buffer and offset for memref operands (#13053)
This is less opinionated than the existing `ukernel.generic` in that
`memref` operands are lowered to just base buffer and offset operands,
the bare minimum for any callee to be able to do anything with the
memref.

By not appending stride values, this leaves it up to each use case to
decide which strides to pass.

This makes `ukernel.generic_raw` flexible enough to replace all other
existing ukernel ops. Though we want to retain `ukernel.generic` for
now, we will at least proceed with migrating `ukernel.mmt4d` and all the
ukernels in VMVXOps.td.

This PR also makes the `offset` be passed along with the buffer in all cases, including the rank-0 case where it was omitted (that was #13054, now folded into this).

This PR also does some naming harmonization: `micro_kernel` ->
`ukernel`, `foo_ukernel` -> `ukernel_foo`.