Adding support for passing processor_id/data to dynamic imports. (#13109)
Adding an `hal.import.fields` attr with string fields to an import will
append those fields on to the call parameter struct. Currently only
`processor_id`/`processor_data` are exposed but we can add support for
more of the environment/dispatch/workgroup state as needed.

Also added support for overriding the import name with `hal.import.name`
to allow for multiple typed imports to call the same imported function.