Add a sample that uses the iree_codegen.ukernel.generic operation to call a microkernel supplied as a plugin. (#13132)
This PR adds a sample that mirrors the standalone_plugin and
system_plugin examples, but instaed of using a stream-level shims
for invoking the micro kernel, uses the iree_codegen.ukernel.generic
to call a specific microkernel. The input IR contains a dispatch
(expressed within flow.dispatch.region) which extracts slices of the
inputs/outputs on a per-workgroup basis and calls the microkernels
provided.