Unify the two plugin examples to use the same micro-kernel. (#13324)
As follow up from #13132 this example uses the same micro-kernel both
as a custom dispatch call and as a micro kernel. This requires adding
a new optional attribute to iree_codegen.ukernel.generic operations
that carry additional attributes to append to the function definition.
Using these hal.import.fields attribute can be set on the function
definitions when lowering to function calls. This is path to send
additional contextual information to the micro-kernel.

Also update the README and op-definition documentation.