Stop trying to skip lint checks if tools are missing (#13411)
This functionality is broken. The DEBUG trap runs *before* each command,
not after. Better to have a clear error message than confusing failures,
which is what we get now. The logic for doing this was also super
verbose and complicated.

We should really rewrite this whole thing in Python.

skip-ci: lint scripts don't affect CI