Remove the >3D restriction on reshaping for dot_general (#13521)
Previously, the collapsing dims together (in case of multiple dims of
the same type) would only happen if the input rank is larger than 3.
This made the incorrect assumption that a 3d tensor is of a standard
form (BxCxP or BxPxC), where in reality this could be the result of
multiples of one dimension but none of another (BxBxC, for example). In
those cases we still need to perform a reshape.