Preserve unsigned integer inputs/outputs (#13090)
Thanks so much to @rsuderman for the guidance and help !

This PR solves a bug where the signedness of inputs and outputs is lost,
because lowering to linalg dialect throws the signedness away. To fix
it, I'm using a new op `tensor.bitcast`, which casts the inputs/outputs
of the function from an unsigned type to the signless type in the body.

I am completely new to IREE, so this PR is still a little rough. Any
critique and suggestions for improvement is highly welcome. I'm
particularly interested in ways to make this more robust.

Fixes #9282
Fixes #12665