Make Discord notification link directly to workflow run (#13646)
This makes the title a URL that takes you to the thing you actually
want to see: the failing run. Currently the closest thing is the checks
page, which is super noisy for commits on main because all the stuff
running on `schedule` or `pull_request_target` run in the context of
the HEAD commit on main and apparently post all their checks there...

I tried also making the message less noisy in general, but it removed
too much context. If we want something different, we'll have to switch
to a different action: this one deliberately already has everything
formatted for you. Although really I'd love to figure out a way we can
use issues for CI failures. Discord kind of sucks for tracking.