Coalescing adjacent stream.async.fills and eliding useless ones. (#13782)
These tend to arise from concats of padding or partially constant values
(`concat(%value, %zero, %zero, %zero)`) and prior to this change would
result in large sequences of fill->fill->fill at runtime. In some cases
a tensor will be created from a splat of a constant (zero, etc) and then
filled with that same constant and now those are handled (when found).
Canonicalization after passes that do more involved analysis
(iree-stream-emplace-allocations and iree-stream-elide-async-writes) do
the heavy lifting and leave the rest of the cleanup to these
canonicalizers.


![image](https://github.com/openxla/iree/assets/75337/995db391-ee3d-4dae-b5de-2ed2470c5ae7)

Fixes #13767.