Switching to --iree-stream-partitioning-favor=min-peak-memory. (#13791)
The current value of max-concurrency can be dangerous on models that are
near the max of device memory and min-peak-memory allows us to still get
some concurrency while being more conservative about extending
lifetimes. Users not concerned with peak memory usage but wanting the
most concurrency can still use the flag to get the old behavior.

Progress on #13545.