[CPU] Consistently set native vector width in LLVMCPUTarget (#13775)
The hardware native vector width was set in LLVMCPUTarget only when we
were able to infer that information from TTI. This led to cases where we
didn't know the native vector width and decided to use a default value
when this information was needed later in the pipeline (i.e., Kernel
Dispatch). On top of that, we had a flag in Kernel Dispatch to let the
user specify a native vector width which was only honored when the
native vector width hadn't been infered from TTI.

This patch simplifies this scenario by *always* setting a native vector
width in LLVMCPUTarget, either the one inferred from TTI, a default one
or one specified by the user. In that way, all the passes now see a
consisten native vector width. We also give priority to the
user-specified vector width so that we can experiment with different
widths regardless of the value inferred by TTI.

A workaround is introduced for the VMVX code in Kernel Dispatch, as VMVX
doesn't go through the LLVM CPU target initialization.