fix the `cpu_ukernel` standalone plugin: was missing `weak.c` (#13850)
I forgot to update this in #13715. This would result in linker errors
when using this outside of x86_64 and arm64. Sorry! @lundong .

(The need for this standalone plugin is likely to diminish as #13804 is
getting completed. Even if it stays around, it could be reimplemented as
a thin shell around the bitcode.).