Unbreak the `byo_llvm.sh` build with `iree_bitcode_library`. (#13968)
The byo-llvm build is broken by the way `iree_bitcode_library` assumes
it's dealing with our own bundled `clang` (#13957). This is an attempt
to fix that. Specifically, this implements the idea in
https://github.com/openxla/iree/issues/13957#issuecomment-1579175544 .

So `iree_bitcode_library` is changed to take `clang`, Clang builtin
headers, and `llvm-link` paths from externally defined variables.

`iree_llvm.cmake` is updated to set these variables appropriately, in
both the bundled and the installed cases.

`byo_llvm.sh` is updated to pass a `LLVM_INSTALL_DIR` to
`iree_llvm.cmake`.

Fixes #13957 .