[LLVMCPU] Add pass to enable Armv9 Streaming SVE mode (#13558)
This patch adds a pass 'iree-llvmcpu-enable-aarch64-ssve' that enables the Armv9 Scalable Matrix Extension (SME) Streaming SVE (SSVE) mode [1].

SSVE is enabled in the LLVM backend at the function boundary by specifying one of the following attributes [2]:

  aarch64_pstate_sm_enabled - calls to functions with this attribute are
                              wrapped with 'smstart sm' / 'smstop sm' [3].
                              Changes the function ABI.

  aarch64_pstate_sm_body    - 'smstart sm' / 'smstop sm' are emitted in the
                              function prologue / epilogue for functions
                              marked with this attribute. This is internal
                              and doesn't change the function ABI.

This pass adds the 'aarch64_pstate_sm_body' attribute to functions via the passthrough mechanism [4].

This attribute is used because PSTATE.SM changes are kept internal to the function and the purpose of this pass is to enable SSVE for dispatch functions which are called by the IREE runtime. The AAPCS64 [5] states it is the caller's responsibility to ensure that PSTATE.SM has a valid value on entry to a callee, the 'aarch64_pstate_sm_enabled' attribute would change the function ABI forcing the caller (IREE runtime) to be responsible for managing PSTATE.SM before entry/exit. At present, the runtime doesn't know the details of dispatches such that it could emit these instructions.

The pass is enabled for AArch64 when SVE(2) and SME are enabled for the following lowering configurations:

  * CPUBufferOpsTileAndVectorize
  * CPUDoubleTilingPeelingExpert
  * CPUConvTileAndDecomposeExpert

These configurations were chosen simply because they're used in one of our pipelines.

[1] https://developer.arm.com/documentation/ddi0616/aa
[2] https://llvm.org/docs/AArch64SME.html
[3] https://developer.arm.com/documentation/ddi0602/2023-03/Base-Instructions/SMSTART--Enables-access-to-Streaming-SVE-mode-and-SME-architectural-state--an-alias-of-MSR--immediate--
[4] https://mlir.llvm.org/docs/Dialects/LLVM/#attribute-pass-through
[5] https://github.com/ARM-software/abi-aa/blob/main/aapcs64/aapcs64.rst#671pstatesm-interfaces